+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [3.5.0] - TBD
+### Fixed
+- Throw exception if `cleanupDelay` can't be parsed instead of returning the default value.
 ### Changed
+- Don't return records for cleanup after 10 attempts have been reached. 
 - Added additional checks for the number of levels in table and partition paths so that invalid paths are not scheduled for deletion.
 
 ## [3.4.14] - 2022-11-25